+++ api/endpoints/subjects/delete.php
@@ -15,14 +15,23 @@ try {
     ]);
 
     $subject = Subject::getByGuid($db, $input->guid);
-    $subject->delete();
+    $books = Book::getBySubject($db, $subject->id);
 
-    Book::deleteBySubject($db, $subject->id);
-
+    foreach ($books as $book) {
+        $asigneeCopies = Book::getCountOfCopiesByBookGuid($db, $book->guid);
+        if ($asigneeCopies != 0)
+            createException("There are " . $asigneeCopies . " students who have copies of any book of this subject");
+        else {
+            $subject->delete();
+            $book->delete();
+        }
+    }
 
     $db->commit();
 
-    Response::sendResponse([]);
+    Response::sendResponse([
+        "data" => true
+    ]);
 } catch (\Exception $th) {
     $db->rollBack();
     print_r(json_encode(array("status" => false, "message" => $th->getMessage(), 'code' => $th->getCode())));
